pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "market-risk-model-validation-toolkit"
version = "0.1.0"
description = "Recruiter-friendly Python toolkit for market risk analytics and model validation workflows."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "matplotlib>=3.8,<4.0",
    "numpy>=1.26,<3.0",
    "pandas>=2.2,<3.0",
    "PyYAML>=6.0,<7.0",
    "scipy>=1.13,<2.0",
    "yfinance>=0.2.54,<1.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-ra"

[tool.ruff]
line-length = 100
target-version = "py310"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "B"]
ignore = ["E501"]
